Setup for get wbz ids.
